Removed 1 unnecessary stubbing in PrerequisitesTaskTest.java
